+++ franka_gazebo/src/franka_hw_sim.cpp
@@ -32,6 +32,7 @@ bool FrankaHWSim::initSim(const std::string& robot_namespace,
   }
 
   this->robot_ = parent;
+  this->efforts_initialized_ = false;
 
 #if GAZEBO_MAJOR_VERSION >= 8
   gazebo::physics::PhysicsEnginePtr physics = gazebo::physics::get_world()->Physics();
@@ -45,6 +46,9 @@ bool FrankaHWSim::initSim(const std::string& robot_namespace,
   // NOTE: Can be overwritten by the user via the 'gravity_vector' ROS parameter.
   auto gravity = physics->World()->Gravity();
   this->gravity_earth_ = {gravity.X(), gravity.Y(), gravity.Z()};
+
+  model_nh.param<double>("tau_ext_lowpass_filter", this->tau_ext_lowpass_filter_,
+                         kDefaultTauExtLowpassFilter);
 
   // Generate a list of franka_gazebo::Joint to store all relevant information
   for (const auto& transmission : transmissions) {
@@ -316,14 +320,15 @@ void FrankaHWSim::writeSim(ros::Time /*time*/, ros::Duration /*period*/) {
 
   for (auto& pair : this->joints_) {
     auto joint = pair.second;
-    auto command = joint->command;
 
     // Check if this joint is affected by gravity compensation
     std::string prefix = this->arm_id_ + "_joint";
     if (pair.first.rfind(prefix, 0) != std::string::npos) {
       int i = std::stoi(pair.first.substr(prefix.size())) - 1;
-      command += g.at(i);
+      joint->gravity = g.at(i);
     }
+
+    auto command = joint->command + joint->gravity;
 
     if (std::isnan(command)) {
       ROS_WARN_STREAM_NAMED("franka_hw_sim",
@@ -475,16 +480,25 @@ void FrankaHWSim::updateRobotState(ros::Time time) {
     this->robot_state_.tau_J[i] = joint->effort;
     this->robot_state_.dtau_J[i] = joint->jerk;
 
-    this->robot_state_.q_d[i] = joint->position;
-    this->robot_state_.dq_d[i] = joint->velocity;
-    this->robot_state_.ddq_d[i] = joint->acceleration;
+    // since we don't support position or velocity interface yet, we set the desired joint
+    // trajectory to zero indicating we are in torque control mode
+    this->robot_state_.q_d[i] = joint->position;  // special case for resetting motion generators
+    this->robot_state_.dq_d[i] = 0;
+    this->robot_state_.ddq_d[i] = 0;
     this->robot_state_.tau_J_d[i] = joint->command;
 
     // For now we assume no flexible joints
     this->robot_state_.theta[i] = joint->position;
     this->robot_state_.dtheta[i] = joint->velocity;
 
-    this->robot_state_.tau_ext_hat_filtered[i] = joint->effort - joint->command;
+    if (this->efforts_initialized_) {
+      double tau_ext = joint->effort - joint->command + joint->gravity;
+
+      // Exponential moving average filter from tau_ext -> tau_ext_hat_filtered
+      this->robot_state_.tau_ext_hat_filtered[i] =
+          this->tau_ext_lowpass_filter_ * tau_ext +
+          (1 - this->tau_ext_lowpass_filter_) * this->robot_state_.tau_ext_hat_filtered[i];
+    }
 
     this->robot_state_.joint_contact[i] = static_cast<double>(joint->isInContact());
     this->robot_state_.joint_collision[i] = static_cast<double>(joint->isInCollision());
@@ -518,6 +532,8 @@ void FrankaHWSim::updateRobotState(ros::Time time) {
   this->robot_state_.control_command_success_rate = 1.0;
   this->robot_state_.time = franka::Duration(time.toNSec() / 1e6 /*ms*/);
   this->robot_state_.O_T_EE = this->model_->pose(franka::Frame::kEndEffector, this->robot_state_);
+
+  this->efforts_initialized_ = true;
 }
 
 }  // namespace franka_gazebo